+++ CHANGELOG.md
@@ -81,6 +81,7 @@
 - Empty attributes are no longer included in the csv export.
 - When Sparv crashes due to unexpected errors, the traceback is now hidden from the user unless the `--log debug`
   argument is used.
+- If the `-j`/`--cores` option is used without an argument, all available CPU cores are used.
 
 ### Fixed